fix(storage): seperate subtitle and fix references
1. Seperate db.ts from index.ts to export discrete methods and SubtitleDatabase.
2. Fix references for SubtitleStorage.
