+++ src/renderer/containers/VideoCanvas.vue
@@ -51,7 +51,7 @@ import { Video as videoActions } from '@/store/actionTypes';
 import { videodata } from '@/store/video';
 import BaseVideoPlayer from '@/components/PlayingView/BaseVideoPlayer.vue';
 import { MediaItem, PlaylistItem } from '../interfaces/IDB';
-import { deleteSubtitlesByPlaylistId } from '../services/storage/SubtitleStorage';
+import { deleteSubtitlesByPlaylistId } from '../services/storage/subtitle';
 
 export default {
   name: 'VideoCanvas',
